MAKE-1331: Make signal processing jobs in cedar report time series to Performance Analysis and Triage Service and remove all other tooling (#230)
