+++ model/change_point_detection.go
@@ -2,6 +2,7 @@ package model
 
 import (
 	"context"
+	"fmt"
 	"math"
 	"time"
 
@@ -109,22 +110,27 @@ func (ts TriageStatus) Validate() error {
 }
 
 type PerformanceResultSeriesID struct {
-	Project string `bson:"project"`
-	Variant string `bson:"variant"`
-	Task    string `bson:"task"`
-	Test    string `bson:"test"`
+	Project   string           `bson:"project"`
+	Variant   string           `bson:"variant"`
+	Task      string           `bson:"task"`
+	Test      string           `bson:"test"`
+	Arguments map[string]int32 `bson:"args"`
+}
+
+func (p PerformanceResultSeriesID) String() string {
+	return fmt.Sprintf("%s %s %s %s %v", p.Project, p.Variant, p.Task, p.Test, p.Arguments)
 }
 
 type TimeSeriesEntry struct {
 	PerfResultID string  `bson:"perf_result_id"`
 	Value        float64 `bson:"value"`
 	Order        int     `bson:"order"`
+	Version      string  `bson:"version"`
 }
 
 type MeasurementData struct {
-	Measurement  string            `bson:"measurement"`
-	TimeSeries   []TimeSeriesEntry `bson:"time_series"`
-	ChangePoints []ChangePoint     `bson:"change_points"`
+	Measurement string            `bson:"measurement"`
+	TimeSeries  []TimeSeriesEntry `bson:"time_series"`
 }
 
 type PerformanceData struct {
@@ -157,7 +163,7 @@ func MarkPerformanceResultsAsAnalyzed(ctx context.Context, env cedar.Environment
 	return nil
 }
 
-func GetPerformanceResultSeriesIdsNeedingChangePointDetection(ctx context.Context, env cedar.Environment) ([]PerformanceResultSeriesID, error) {
+func GetPerformanceResultSeriesIdsNeedingTimeSeriesUpdate(ctx context.Context, env cedar.Environment) ([]PerformanceResultSeriesID, error) {
 	cur, err := env.GetDB().Collection(perfResultCollection).Aggregate(ctx, []bson.M{
 		{
 			"$match": bson.M{
@@ -229,6 +235,7 @@ func GetPerformanceResultSeriesIDs(ctx context.Context, env cedar.Environment) (
 					"variant": "$" + bsonutil.GetDottedKeyName(perfInfoKey, perfResultInfoVariantKey),
 					"task":    "$" + bsonutil.GetDottedKeyName(perfInfoKey, perfResultInfoTaskNameKey),
 					"test":    "$" + bsonutil.GetDottedKeyName(perfInfoKey, perfResultInfoTestNameKey),
+					"args":    "$" + bsonutil.GetDottedKeyName(perfInfoKey, perfResultInfoArgumentsKey),
 				},
 			},
 		},
@@ -264,20 +271,6 @@ func GetPerformanceData(ctx context.Context, env cedar.Environment, performanceR
 		{
 			"$unwind": "$" + bsonutil.GetDottedKeyName(perfRollupsKey, perfRollupsStatsKey),
 		},
-		//Filter out any change points unrelated to this rollup
-		{
-			"$addFields": bson.M{
-				bsonutil.GetDottedKeyName(perfAnalysisKey, perfAnalysisChangePointsKey): bson.M{
-					"$filter": bson.M{
-						"input": "$" + bsonutil.GetDottedKeyName(perfAnalysisKey, perfAnalysisChangePointsKey),
-						"as":    "cp",
-						"cond": bson.M{
-							"$eq": bson.A{"$$cp.measurement", "$" + bsonutil.GetDottedKeyName(perfRollupsKey, perfRollupsStatsKey, perfRollupValueNameKey)},
-						},
-					},
-				},
-			},
-		},
 		{
 			"$group": bson.M{
 				"_id": bson.M{
@@ -286,6 +279,7 @@ func GetPerformanceData(ctx context.Context, env cedar.Environment, performanceR
 					"task":        "$" + bsonutil.GetDottedKeyName(perfInfoKey, perfResultInfoTaskNameKey),
 					"test":        "$" + bsonutil.GetDottedKeyName(perfInfoKey, perfResultInfoTestNameKey),
 					"measurement": "$" + bsonutil.GetDottedKeyName(perfRollupsKey, perfRollupsStatsKey, perfRollupValueNameKey),
+					"args":        "$" + bsonutil.GetDottedKeyName(perfInfoKey, perfResultInfoArgumentsKey),
 				},
 				"time_series": bson.M{
 					"$push": bson.M{
@@ -297,21 +291,7 @@ func GetPerformanceData(ctx context.Context, env cedar.Environment, performanceR
 						},
 						"order":          "$" + bsonutil.GetDottedKeyName(perfInfoKey, perfResultInfoOrderKey),
 						"perf_result_id": "$_id",
-					},
-				},
-				"change_points": bson.M{
-					"$push": "$" + bsonutil.GetDottedKeyName(perfAnalysisKey, perfAnalysisChangePointsKey),
-				},
-			},
-		},
-		// Flatten the change points into one array
-		{
-			"$addFields": bson.M{
-				"change_points": bson.M{
-					"$reduce": bson.M{
-						"input":        "$change_points",
-						"initialValue": bson.A{},
-						"in":           bson.M{"$concatArrays": bson.A{"$$value", "$$this"}},
+						"version":        "$" + bsonutil.GetDottedKeyName(perfInfoKey, perfResultInfoVersionKey),
 					},
 				},
 			},
@@ -323,12 +303,12 @@ func GetPerformanceData(ctx context.Context, env cedar.Environment, performanceR
 					"variant": "$_id.variant",
 					"task":    "$_id.task",
 					"test":    "$_id.test",
+					"args":    "$_id.args",
 				},
 				"data": bson.M{
 					"$push": bson.M{
-						"measurement":   "$_id.measurement",
-						"time_series":   "$time_series",
-						"change_points": "$change_points",
+						"measurement": "$_id.measurement",
+						"time_series": "$time_series",
 					},
 				},
 			},
@@ -353,7 +333,7 @@ func GetPerformanceData(ctx context.Context, env cedar.Environment, performanceR
 func ReplaceChangePoints(ctx context.Context, env cedar.Environment, performanceData *PerformanceData, mappedChangePoints map[string][]ChangePoint) error {
 	err := clearUntriagedChangePoints(ctx, env, performanceData.PerformanceResultId)
 	if err != nil {
-		return errors.Wrapf(err, "Unable to clear change points for measurement %s", performanceData.PerformanceResultId)
+		return errors.Wrapf(err, "Unable to clear change points for measurement %s", performanceData.PerformanceResultId.String())
 	}
 	catcher := grip.NewBasicCatcher()
 	for _, measurementData := range performanceData.Data {
